requirements.txt
argparse==1.4.0
requests==2.27.1
python-telegram-bot==13.11
python-dotenv==0.20.0
Pillow==9.1.0
moviepy==1.0.3